Authorization Service
I am updating the Authorization service, and adding testing functionality to it. The first test is just to make sure that the testing actually works. The authenticator is going to be changed to work in tandem with the Session Manger, so functionalities like (authenticate) are going to be taken away from the authService, and managed by the SessionManager through the authService API